Make request IDs 64 bit in PermissionBubbleMediaAccessHandler
This is expected to prevent potential overflows.

(cherry picked from commit b29987b9d50e085a5aefc8d48ad3e4446aaee51e)

Bug: 1178032
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2797413
Auto-Submit: Guido Urdaneta <guidou@chromium.org>
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#868216}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2822148
Commit-Queue: Jana Grill <janagrill@chromium.org>
Cr-Commit-Position: refs/branch-heads/4240@{#1599}
Cr-Branched-From: f297677702651916bbf65e59c0d4bbd4ce57d1ee-refs/heads/master@{#800218}

+++ chrome/browser/media/webrtc/permission_bubble_media_access_handler.cc
@@ -229,7 +229,7 @@ void PermissionBubbleMediaAccessHandler::ProcessQueuedAccessRequest(
 
   DCHECK(!it->second.empty());
 
-  const int request_id = it->second.begin()->first;
+  const int64_t request_id = it->second.begin()->first;
   const content::MediaStreamRequest& request =
       it->second.begin()->second.request;
 #if defined(OS_ANDROID)
@@ -290,7 +290,7 @@ void PermissionBubbleMediaAccessHandler::RegisterProfilePrefs(
 
 void PermissionBubbleMediaAccessHandler::OnMediaStreamRequestResponse(
     content::WebContents* web_contents,
-    int request_id,
+    int64_t request_id,
     content::MediaStreamRequest request,
     const blink::MediaStreamDevices& devices,
     blink::mojom::MediaStreamRequestResult result,
@@ -322,7 +322,7 @@ void PermissionBubbleMediaAccessHandler::OnMediaStreamRequestResponse(
 
 void PermissionBubbleMediaAccessHandler::OnAccessRequestResponse(
     content::WebContents* web_contents,
-    int request_id,
+    int64_t request_id,
     const blink::MediaStreamDevices& devices,
     blink::mojom::MediaStreamRequestResult result,
     std::unique_ptr<content::MediaStreamUI> ui) {
